Add files via upload
My scratch project is about a fish being alone and scared. The villain in this story is a shark that looks to eat the fish, a 30 second sprint of avoiding the shark will get you a win. If the shark gets you then game over and try again. Click on the flag for quick dialogue, the space bar to start the game and the up arrow and down arrow to move the hero of the story up or down to avoid the shark